+++ index.html
@@ -8,7 +8,8 @@
     <title>Login Books Riwi</title>
   </head>
   <body>
-    <header>
+    <div id="root"></div>
+    <!-- <header>
 
     </header>
     <main>
@@ -29,7 +30,7 @@
     </main>
     <footer>
 
-    </footer>
+    </footer> -->
     <script type="module" src="./src/index.ts"></script>
   </body>
 </html>
